backend/requirements.txt
# ==================================
# CONTROLNOT V2 - PYTHON DEPENDENCIES
# ==================================

# Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI Providers
openai==1.10.0
httpx==0.26.0

# Google Cloud (OCR + Drive)
google-cloud-vision==3.5.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Document Generation
python-docx==1.1.0

# Utils
python-dotenv==1.0.0
structlog==24.1.0
aiofiles==23.2.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Date handling
python-dateutil==2.8.2

# Testing (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
